+++ src/vue/modules/create-sale-form/components/information-step-form.vue
@@ -363,7 +363,9 @@ export default {
         this.form.quoteAssets = []
       }
 
-      this.availableQuoteAssets = await this.loadBaseAssetsByQuote(value)
+      const quoteAssets = await this.loadBaseAssetsByQuote(value)
+
+      this.availableQuoteAssets = [this.form.capAsset, ...quoteAssets]
       this.isQuoteAssetsLoaded = true
     },
   },
@@ -391,9 +393,6 @@ export default {
     },
 
     populateForm () {
-      const quoteAssets = this.request.quoteAssets
-        .filter(quoteAsset => this.request.defaultQuoteAsset !== quoteAsset)
-
       this.form.name = this.request.name
       this.form.baseAsset = this.ownedAssets
         .find(item => item.code === this.request.baseAsset)
@@ -404,7 +403,7 @@ export default {
       this.form.softCap = this.request.softCap
       this.form.hardCap = this.request.hardCap
       this.form.assetsToSell = this.request.assetsToSell
-      this.form.quoteAssets = quoteAssets
+      this.form.quoteAssets = this.request.quoteAssets
       this.form.isWhitelisted = this.request.isWhitelisted
     },